提交

+++ build.sh
@@ -1,9 +1,18 @@
 #!/bin/bash
-meteor build --server https://cn.steedos.com/records --directory /srv/creator_records --allow-superuser
-cd /srv/creator_records/bundle/programs/server
-rm -rf node_modules
-rm -f npm-shrinkwrap.json
-npm install --registry https://registry.npm.taobao.org -d
+BUNDLE_PATH="/srv/creator_records"
 
-cd /srv/creator_records/
-pm2 restart creator_records.0
+# Create BUNDLE_PATH directory if it does not exist
+[ ! -d $BUNDLE_PATH ] && mkdir -p $BUNDLE_PATH || :
+
+if [ -d "$BUNDLE_PATH" ]; then
+	meteor build --server https://cn.steedos.com/creator_records --directory $BUNDLE_PATH --allow-superuser
+	cd $BUNDLE_PATH/bundle/programs/server
+	rm -rf node_modules
+	rm -f npm-shrinkwrap.json
+	npm install --registry https://registry.npm.taobao.org -d
+
+	cd $BUNDLE_PATH
+	pm2 restart creator_records.0
+else
+	echo "!!!=> Failed to create bundle path: $BUNDLE_PATH"
+fi
